only do the linking if its not already linked up

+++ install.sh
@@ -48,5 +48,9 @@ function add_bundle(){
 add_bundle "nerdtree"
 add_bundle "camelcasemotion"
 add_bundle "surround"
-ln -s "$current_dir/oh-my-zsh" "$HOME/.oh-my-zsh"
-ln -s "$current_dir/oh-my-zsh-custom" "$HOME/.oh-my-zsh-custom"
+if [ ! -e  "$HOME/.oh-my-zsh" ]; then
+    ln -s "$current_dir/oh-my-zsh" "$HOME/.oh-my-zsh"
+fi
+if [ ! -e  "$HOME/.oh-my-zsh-custom" ]; then
+    ln -s "$current_dir/oh-my-zsh-custom" "$HOME/.oh-my-zsh-custom"
+fi
